third_party/dawn/third_party/catapult/third_party/vinn/third_party/parse5/test/data/tree_construction_regression/gh40_form_in_template.dat

#data
<template><form><input name="q"></form><div>second</div></template>
#document-fragment
template
#errors
#document
| <template>
|   content
|     <form>
|       <input>
|         name="q"
|     <div>
|       "second"
